fix cae_pt for py310

+++ tasks/ssl/cae/util/masking_generator.py
@@ -76,7 +76,7 @@ class MaskingGenerator:
         return delta
 
     def __call__(self):
-        mask = np.zeros(shape=self.get_shape(), dtype=np.int)
+        mask = np.zeros(shape=self.get_shape(), dtype=np.int32)
         mask_count = 0
         while mask_count != self.num_masking_patches:
             max_mask_patches = self.num_masking_patches - mask_count
